change sizes to size_t

+++ src/squid.h
@@ -351,7 +351,7 @@ extern char ThisCache[];	/* main.c */
 
 extern int objcacheStart _PARAMS((int, const char *, StoreEntry *));
 extern void send_announce _PARAMS((void *unused));
-extern int sslStart _PARAMS((int fd, const char *, request_t *, char *, int *sz));
+extern int sslStart _PARAMS((int fd, const char *, request_t *, char *, size_t *sz));
 extern const char *storeToString _PARAMS((const StoreEntry *));
 extern int waisStart _PARAMS((int, const char *, method_t, char *, StoreEntry *));
 extern void storeDirClean _PARAMS((void *unused));
@@ -360,7 +360,7 @@ extern int passStart _PARAMS((int fd,
 	request_t * request,
 	char *buf,
 	int buflen,
-	int *size_ptr));
+	size_t *size_ptr));
 extern void identStart _PARAMS((int, icpStateData *,
 	void       (*callback) _PARAMS((void *))));
